+++ Sidus_Github_Solution/Drone_Flight_Controller_2/Drone_Flight_Controller_2.vcxproj
@@ -63,10 +63,10 @@
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\WiFi\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\Wire\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\ArduinoOTA\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\Update\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\ESPmDNS\src;$(ProjectDir)..\..\..\..\Arduino\libraries;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries;E:\Arduino Sketches\libraries;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\cores\esp32;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\cores\esp32\libb64;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\variants\esp32;$(ProjectDir)..\Drone_Flight_Controller_2;C:\Users\ahmet\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include\c++\4.8.2;C:\Users\ahmet\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include\c++\4.8.2\xtensa-lx106-elf;C:\Users\ahmet\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include;C:\Users\ahmet\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\lib\gcc\xtensa-lx106-elf\4.8.2\include;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\WiFi\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\Wire\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\ArduinoOTA\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\Update\src;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries\ESPmDNS\src;$(ProjectDir)..\..\..\..\Arduino\libraries;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\libraries;$(ProjectDir)..\..\..\..\..\..\OneDrive\Documents\Arduino\libraries;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\cores\esp32;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\cores\esp32\libb64;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\variants\esp32;$(ProjectDir)..\Drone_Flight_Controller_2;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\config;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\bluedroid;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\app_update;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\bootloader_support;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\bt;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\driver;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\esp32;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\ethernet;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\fatfs;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\freertos;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\jsmn;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\log;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\mdns;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\mbedtls;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\mbedtls_port;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\newlib;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\nvs_flash;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\openssl;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\soc;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\spi_flash;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\sdmmc;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\tcpip_adapter;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\ulp;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\vfs;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\wear_levelling;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\xtensa-debug-module;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\coap;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\wpa_supplicant;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\expat;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\json;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\nghttp;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include\lwip;$(ProjectDir)..\..\..\..\..\..\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include\c++\4.8.2;$(ProjectDir)..\..\..\..\..\..\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include\c++\4.8.2\xtensa-lx106-elf;$(ProjectDir)..\..\..\..\..\..\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\xtensa-lx106-elf\include;$(ProjectDir)..\..\..\..\..\..\AppData\Local\arduino15\packages\esp8266\tools\xtensa-lx106-elf-gcc\1.20.0-26-gb404fb9-2\lib\gcc\xtensa-lx106-elf\4.8.2\include;$(ProjectDir)..\..\..\..\Arduino\hardware\espressif\esp32\tools\sdk\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <ForcedIncludeFiles>$(ProjectDir)__vm\.Drone_Flight_Controller_2.vsarduino.h;%(ForcedIncludeFiles)</ForcedIncludeFiles>
       <WholeProgramOptimization>false</WholeProgramOptimization>
-      <PreprocessorDefinitions>__ESP32_ESp32__;__ESP32_ESP32__;ESP_PLATFORM;HAVE_CONFIG_H;F_CPU=240000000L;ARDUINO=10802;ARDUINO_ESP32_DEV;ARDUINO_ARCH_ESP32;ESP32;CORE_DEBUG_LEVEL=0;__cplusplus=201103L;_VMICRO_INTELLISENSE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>__ESP32_ESp32__;__ESP32_ESP32__;ESP_PLATFORM;HAVE_CONFIG_H;F_CPU=240000000L;ARDUINO=10803;ARDUINO_ESP32_DEV;ARDUINO_ARCH_ESP32;ESP32;CORE_DEBUG_LEVEL=0;__cplusplus=201103L;_VMICRO_INTELLISENSE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
@@ -115,7 +115,6 @@
     <ClCompile Include="cMsgUdpR01.cpp" />
     <ClCompile Include="cMsgUdpT01.cpp" />
     <ClCompile Include="cRxFilter.cpp" />
-    <ClCompile Include="dataFilter.cpp" />
     <ClCompile Include="kalmanFilter.cpp" />
     <ClCompile Include="kalmanFilter_initialize.cpp" />
     <ClCompile Include="Local_HMC5883L.cpp" />
