+++ Sidus_Github_Solution/Drone_Flight_Controller_2/__vm/.Drone_Flight_Controller_2.vsarduino.h
@@ -17,7 +17,7 @@
 #define ESP_PLATFORM
 #define HAVE_CONFIG_H
 #define F_CPU 240000000L
-#define ARDUINO 10802
+#define ARDUINO 10803
 #define ARDUINO_ESP32_DEV
 #define ARDUINO_ARCH_ESP32
 #define ESP32
